Move generic views to base_views.py
